+++ contracts/os/andromeda-kernel/src/contract.rs
@@ -176,6 +176,9 @@ pub fn query(deps: Deps, _env: Env, msg: QueryMsg) -> Result<Binary, ContractErr
         QueryMsg::ChainNameByChannel { channel } => {
             encode_binary(&query::chain_name_by_channel(deps, channel)?)
         }
+        QueryMsg::PendingPackets { channel_id } => {
+            encode_binary(&query::pending_packets(deps, channel_id)?)
+        }
         QueryMsg::GetEnv { variable } => encode_binary(&query::get_env(deps, variable)?),
     }
 }